requirements.txt
pyvirtualdisplay
celery
eventlet
Flask
Flask-Cors
Flask-Login
Flask-SQLAlchemy
Flask-redis
lxml
parsel
PyExecJS
PyMySQL
requests
tenacity
WTForms
gunicorn
opencv-python
numpy
beautifulsoup4
pillow
selenium
xgboost
scikit-learn